Start the spherex document class
This custom class implements the format for pipeline module
descriptions, but as a reusable class.

The document type can be specified as a class option and that provides a
way to modify the appearance of different classes of documents.

The test/ directory contains several test documents that exercise
different document settings.