Add Docker image build
This docker build is largely based on the system existing in
github.com/lsst/lsst-texmf. The Docker image is based on the Python
official image, which makes it convenient to run Python scripts as part
of a document's build.

The Docker installation is managed by the
docker/install-base-packages.sh script. Additional apt-get packages can
be installed by updating that script. Using an install script in
conjunction with a build stage helps to minimize the Docker image size.