Run tests inside the Docker container
For this to work, I need to run a local docker registry to be able to
upload the image from one step and use it in the next.

I also found I needed to run tests in docker from root. This ensures
that the repo's .git/ directory is included in the mounted workspace so
that meta.tex is rendered correctly.