Add \dochistory and \addtohist commands
dochistory is an environment that formats a longtable with version,
version, date, description, and author columns.

The \addtohist command adds a single row to this table.

This is based on the lsstdoc and gaia classes.

The `test/pipeline_module` test case includes the dochistory command as
a demonstration.

The README includes some documentation of this commands.