Disable git safe directory checks
The /workspace directory that gets mounted into the container may not
have the expected ownership, and this can result in errors with Git's
new safe directory checks. See

https://git-scm.com/docs/git-config/2.35.2#Documentation/git-config.txt-safedirectory

This new git command in the Docker file adds a git configuration to the
container that effectively disables these checks. This is fine because
of the nature of the isolated container.